feat(ExceptionHandler): rename class and add case format
